refactor: code related to moving window on macOS
