+++ lib/main.dart
@@ -39,7 +39,6 @@ import 'config/shortcuts.dart';
 import 'widgets/router/no_effect_page_route.dart';
 import 'widgets/title_bar/window_frame.dart';
 import 'widgets/shortcuts/router_actions_manager.dart';
-import 'widgets/navigation_bar/utils/macos_move_window.dart';
 import 'widgets/ax_reveal/widgets/reveal_effect_context.dart';
 import 'widgets/router/rune_with_navigation_bar_and_playback_controllor.dart';
 
@@ -309,23 +308,14 @@ class _RuneState extends State<Rune> {
           onGenerateRoute: (settings) {
             final routeName = settings.name!;
 
-            if (!Platform.isMacOS) {
-              if (routeName == '/' || routeName == '/scanning') {
-                return NoEffectPageRoute<dynamic>(
-                  settings: settings,
-                  builder: (context) =>
-                      WindowFrame(routes[routeName]!(context)),
-                );
-              }
-            } else {
-              if (routeName == '/' || routeName == '/scanning') {
-                return NoEffectPageRoute<dynamic>(
-                  settings: settings,
-                  builder: (context) => MacOSMoveWindow(
-                    child: WindowFrame(routes[routeName]!(context)),
-                  ),
-                );
-              }
+            if (routeName == '/' || routeName == '/scanning') {
+              return NoEffectPageRoute<dynamic>(
+                settings: settings,
+                builder: (context) => WindowFrame(
+                  routes[routeName]!(context),
+                  customRouteName: routeName,
+                ),
+              );
             }
 
             final page = RuneWithNavigationBarAndPlaybackControllor(
