+++ lib/widgets/navigation_bar/navigation_bar.dart
@@ -136,35 +136,6 @@ class NavigationBarState extends State<NavigationBar> {
                   : Container();
             });
 
-        late Widget macOSParentWidget = SmallerOrEqualTo(
-            deviceType: DeviceType.fish,
-            builder: (context, isFish) {
-              final movableParentWidget = MacOSMoveWindow(
-                  isEnabledDoubleTap: false, child: parentWidget);
-
-              if (isFish) {
-                return MacOSMoveWindow(
-                    isEnabledDoubleTap: false, child: movableParentWidget);
-              }
-
-              final MoveWindow = MacOSMoveWindow(
-                child: Container(
-                  height: 80,
-                ),
-              );
-
-              return parent != null
-                  ? IntrinsicHeight(
-                      child: Row(
-                        children: [
-                          movableParentWidget,
-                          Expanded(child: MoveWindow)
-                        ],
-                      ),
-                    )
-                  : IntrinsicHeight(child: MoveWindow);
-            });
-
         final baseSlibings = (slibings ?? emptySlibings);
         final validSlibings = isZune
             ? baseSlibings
@@ -211,50 +182,7 @@ class NavigationBarState extends State<NavigationBar> {
           },
         );
 
-        final macOSChildrenWidget = SmallerOrEqualTo(
-          deviceType: DeviceType.fish,
-          builder: (context, isFish) {
-            final movableChildrenWidget = MacOSMoveWindow(
-              isEnabledDoubleTap: false,
-              child: childrenWidget,
-            );
-
-            final MoveWindow = MacOSMoveWindow();
-
-            return parent != null
-                ? IntrinsicHeight(
-                    child: Row(
-                      children: [
-                        movableChildrenWidget,
-                        Expanded(child: MoveWindow)
-                      ],
-                    ),
-                  )
-                : IntrinsicHeight(
-                    child: Row(
-                      children: [
-                        Container(
-                          width: 0,
-                          child: movableChildrenWidget,
-                        ),
-                        Expanded(child: MoveWindow)
-                      ],
-                    ),
-                  );
-          },
-        );
-
         final isSearch = path == '/search';
-
-        final navigationContent = Platform.isMacOS
-            ? [
-                macOSParentWidget,
-                macOSChildrenWidget,
-              ]
-            : [
-                parentWidget,
-                childrenWidget,
-              ];
 
         return Stack(
           children: [
@@ -269,7 +197,10 @@ class NavigationBarState extends State<NavigationBar> {
                     child: Column(
                       crossAxisAlignment: CrossAxisAlignment.stretch,
                       mainAxisSize: MainAxisSize.max,
-                      children: navigationContent,
+                      children: [
+                        parentWidget,
+                        childrenWidget
+                      ],
                     ),
                   ),
                 ),
